src/requirements.txt
pandas
numpy
Flask
Flask-Cors
psycopg2-binary
SQLAlchemy
flask-sqlalchemy

